Make vote function more modular
In this commit, we transfer responsibility of extracting the relevant
data that the vote function will need to act on to the reducer. That way
we know that the only data that the vote function gets is the relevant
data. This allows us to scale our application more easily and ensures
that each function is only getting the information it needs.